pyproject.toml
[tool.poetry]
name = "electron-py-example"
version = "0.1.0"
description = "Connecting Python 3 and Electron example"
authors = ["Cristian Javier Azulay <cjadeveloper@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
simplecalculator = "^0.0.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.5.2"
pre-commit = "^2.7.1"
pytest = "^5.2"
flakehell = "^0.6.1"
pylint = "^2.6.0"
ipython = "^7.18.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
